+++ jdaviz/core/user_api.py
@@ -3,7 +3,8 @@ import astropy.units as u
 
 __all__ = ['UserApiWrapper', 'PluginUserApi', 'ViewerUserApi']
 
-_internal_attrs = ('_obj', '_expose', '_items', '_readonly', '__doc__', '_deprecation_msg')
+_internal_attrs = ('_obj', '_expose', '_items', '_readonly', '_exclude_from_dict',
+                   '__doc__', '_deprecation_msg')
 
 
 class UserApiWrapper:
@@ -11,10 +12,11 @@ class UserApiWrapper:
     This is an API wrapper around an internal object.  For a full list of attributes/methods,
     call dir(object).
     """
-    def __init__(self, obj, expose=[], readonly=[]):
+    def __init__(self, obj, expose=[], readonly=[], exclude_from_dict=[]):
         self._obj = obj
         self._expose = list(expose) + list(readonly)
         self._readonly = readonly
+        self._exclude_from_dict = exclude_from_dict
         self._deprecation_msg = None
         if obj.__doc__ is not None:
             self.__doc__ = self.__doc__ + "\n\n\n" + obj.__doc__
@@ -89,6 +91,32 @@ class UserApiWrapper:
             except AttributeError:
                 continue
 
+    def to_dict(self):
+        def _value(item):
+            if hasattr(item, 'to_dict'):
+                return _value(item.to_dict())
+            if hasattr(item, 'selected'):
+                return item.selected
+            return item
+
+        return {k: _value(getattr(self, k)) for k in self._expose
+                if k not in ('show_api_hints', 'keep_active')
+                and k not in self._exclude_from_dict
+                and not hasattr(getattr(self, k), '__call__')}
+
+    def from_dict(self, d):
+        # loop through expose so that plugins can dictate the order that items should be populated
+        for k in self._expose:
+            if k not in d:
+                continue
+            v = d.get(k)
+            if hasattr(getattr(self, k), '__call__'):
+                raise ValueError(f"cannot overwrite callable {k}")
+            if hasattr(getattr(self, k), 'from_dict') and isinstance(v, dict):
+                getattr(self, k).from_dict(v)
+            else:
+                setattr(self, k, v)
+
 
 class PluginUserApi(UserApiWrapper):
     """
@@ -99,12 +127,12 @@ class PluginUserApi(UserApiWrapper):
     For example::
       help(plugin_object.show)
     """
-    def __init__(self, plugin, expose=[], readonly=[]):
+    def __init__(self, plugin, expose=[], readonly=[], excl_from_dict=[]):
         expose = list(set(list(expose) + ['open_in_tray', 'close_in_tray', 'show']))
         if plugin.uses_active_status:
             expose += ['keep_active', 'as_active']
         self._deprecation_msg = None
-        super().__init__(plugin, expose, readonly)
+        super().__init__(plugin, expose, readonly, excl_from_dict)
 
     def __repr__(self):
         if self._deprecation_msg:
@@ -122,9 +150,9 @@ class ViewerUserApi(UserApiWrapper):
     For example::
       help(viewer_object.show)
     """
-    def __init__(self, viewer, expose=[], readonly=[]):
+    def __init__(self, viewer, expose=[], readonly=[], excl_from_dict=[]):
         expose = list(set(list(expose) + []))
-        super().__init__(viewer, expose, readonly)
+        super().__init__(viewer, expose, readonly, excl_from_dict)
 
     def __repr__(self):
         return f'<{self._obj.reference} API>'